pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlox-openclaw-plugin"
version = "0.1.0"
description = "OpenClaw external plugin for MLOX"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "MLOX Contributors"}
]
dependencies = [
    "busysloths-mlox>=0.1.0"
]

[project.urls]
Homepage = "https://github.com/BusySloths/mlox-openclaw-plugin"
Repository = "https://github.com/BusySloths/mlox-openclaw-plugin"

[project.entry-points."mlox.service_plugins"]
openclaw-native = "mlox_openclaw_plugin.plugin:service_plugin"
